# docs/swagger.yaml
basePath: /
definitions:
...
  requests.CreateUserRequest:
    properties:
      bio:
        example: よろしくお願いします！
        type: string
      birth_date:
        example: "2000-01-01"
        type: string
      display_name:
        example: 山田太郎
        type: string
      gender:
        example: male
        type: string
      profile_image_base64:
        example: data:image/jpeg;base64,...
        type: string
    type: object
...
  requests.UpdateUserRequest:
    properties:
      age:
        example: 25
        type: integer
      bio:
        example: よろしくお願いします！
        type: string
      display_name:
        example: 山田太郎
        type: string
      gender:
        example: male
        type: string
      profile_image_url:
        example: https://example.com/images/profile.jpg
        type: string
    type: object
...
  response.AvatarAI:
    properties:
      bio:
        example: あなたの性格や好みを反映した分身AIです
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FDV
        type: string
      name:
        example: 私の分身AI
        type: string
      personality:
        additionalProperties:
          type: string
        example:
          ' "style"': ' "casual"}'
          '{"tone"': ' "friendly"'
        type: object
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: Firebase IDトークンが不正または無効です
        type: string
    type: object
  response.GetAvatarAIResponse:
    properties:
      avatar_ai:
        $ref: '#/definitions/response.AvatarAI'
    type: object
...
  response.GetMeResponse:
    properties:
      is_registered:
        example: true
        type: boolean
      user:
        $ref: '#/definitions/response.User'
    type: object
...
  response.GetUserAvatarAIResponse:
    properties:
      avatar_ai:
        $ref: '#/definitions/response.AvatarAI'
    type: object
...
  response.GetUserDetailResponse:
    properties:
      user:
        $ref: '#/definitions/response.UserDetail'
    type: object
...
  response.User:
    properties:
      age:
        example: 25
        type: integer
      bio:
        example: よろしくお願いします！
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      display_name:
        example: 山田太郎
        type: string
      gender:
        example: male
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      onboarding_completed:
        example: false
        type: boolean
      profile_image_url:
        example: https://example.com/images/profile.jpg
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  response.UserDetail:
    properties:
      age:
        example: 24
        type: integer
      avatar_ai_accessible:
        description: チャット開始済みかどうか
        example: false
        type: boolean
      bio:
        example: よろしくお願いします！
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      display_name:
        example: 佐藤花子
        type: string
      gender:
        example: female
        type: string
      has_avatar_ai:
        example: true
        type: boolean
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      interests:
        example:
        - '["読書"'
        - ' "映画"]'
        items:
          type: string
        type: array
      location:
        example: 東京都
        type: string
      occupation:
        example: エンジニア
        type: string
      profile_image_url:
        example: https://example.com/images/profile2.jpg
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
  title: Hackathon API
  version: "1.0"
paths:
...
  /users:
    post:
      description: ユーザーを作成する
      parameters:
      - description: ユーザー作成リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      responses:
        "200":
          description: ユーザー作成成功
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: ユーザー作成
      tags:
      - users
  /users/{userId}:
    get:
      description: 特定ユーザーの公開情報を取得する（段階的に情報が公開される）
      parameters:
      - description: ユーザーID（ULID）
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: ユーザー情報取得成功
          schema:
            $ref: '#/definitions/response.GetUserDetailResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: ユーザーが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 特定ユーザーの公開情報取得
      tags:
      - users
  /users/{userId}/avatar-ai:
    get:
      description: 特定ユーザーの分身AI情報を取得する（チャット開始後のみアクセス可能）
      parameters:
      - description: ユーザーID（ULID）
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: 分身AI情報取得成功
          schema:
            $ref: '#/definitions/response.GetUserAvatarAIResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: チャットが開始されていないためアクセス不可
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: ユーザーまたは分身AIが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 特定ユーザーの分身AI取得
      tags:
      - users
...
  /users/me:
    get:
      description: 自分の全情報を取得する（マイページ用）。ユーザーが登録されていない場合はis_registered=falseを返す。
      responses:
        "200":
          description: 自分の情報取得成功
          schema:
            $ref: '#/definitions/response.GetMeResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 自分の情報取得（マイページ用）
      tags:
      - users
    put:
      description: 自分のプロフィール情報を更新する
      parameters:
      - description: ユーザー更新リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateUserRequest'
      responses:
        "200":
          description: 情報更新成功
          schema:
            $ref: '#/definitions/response.User'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 自分の情報更新
      tags:
      - users
  /users/me/avatar-ai:
    get:
      description: 自分の分身AI情報を取得する
      responses:
        "200":
          description: 分身AI情報取得成功
          schema:
            $ref: '#/definitions/response.GetAvatarAIResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 分身AIが存在しない（オンボーディング未完了）
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 自分の分身AI情報取得
      tags:
      - users
...
securityDefinitions:
  Bearer:
    description: 'Firebase IDトークンをBearerトークンとして送信してください。例: "Bearer eyJhbGciOiJSUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"